% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbssr.R
\name{bbssr}
\alias{bbssr}
\title{Block bootstrapped Spearman's Rank Correlation Trend Test}
\usage{
bbssr(x, ci = 0.95, nsim = 2000, eta = 1, bl.len = NULL)
}
\arguments{
\item{x}{- Time series data vector}

\item{ci}{- Confidence Interval}

\item{nsim}{- Number of simulations}

\item{eta}{- Added to the block length}

\item{bl.len}{- Block length}
}
\value{
SR.orig- Spearman's Correlation Coefficient

SRtsrc- Test Statistic

lb- Lower bound of confidence interval value

ub- Upper bound of confidence interval value
}
\description{
Significant serial correlation present in time series data can be accounted for using the nonparametric block bootstrap technique, which incorporates the Mann-Kendall trend test (Kundzewicz and Robson, 2000).  Predetermined block lengths are used in resampling the original time series, thus retaining the memory structure of the data.  If the value of the test statistic falls in the tails of the empirical bootstrapped distribution, there is likely a trend in the data.  The block bootstrap technique is powerful in the presence of auto-correlation (Khaliq et al. 2009; Önöz and Bayazit, 2012).
}
\details{
Block lengths are automatically selected using the lag of the least significance serial correlation, to which the 'eta'  term is added. A value of eta = 1 is used as the default as per Khaliq et al. (2009).  Alternatively, the user may define the block length.  2000 bootstrap replicates are recommended as per Svensson et al. (2005) and Önöz, B. and Bayazit (2012).
}
\examples{
x<-c(Nile)
bbssr(x)

}
\references{
Box, G. E. P. and Jenkins, G. M. (1970). Time Series Analysis Forecasting and Control. Holden-Day, San Fransisco, California, 712 pp.

Khaliq, M. N., Ouarda, T. B. M. J., Gachon, P., Sushama, L., and St-Hilaire, A. (2009). Identification of hydrological trends in the presence of serial and cross correlations: A review of selected methods and their application to annual flow regimes of Canadian rivers. Journal of Hydrology, 368: 117-130.

Kundzewicz, Z. W. and Robson, A. J. (2000). Detecting Trend and Other Changes in Meteorological Data. World Climate Program ? Data and Monitoring. World Meteorological Organization, Geneva  (WMO/TD-No. 1013).

Kundzewicz, Z. W. and Robson, A. J. (2004). Change detection in hydrological records ? a review of the methodology. Hydrological Sciences Journal, 49(1): 7-19.

Lehmann, E. L. (1975). Nonparametrics: statistical methods based on ranks. Holden-Day, Inc., California, 457 pp.

Önöz, B. and Bayazit M. (2012). Block bootstrap for Mann-Kendall trend test of serially dependent data. Hydrological Processes, 26: 3552-3560.

Sneyers, R. (1990). On the statistical analysis of series of observations. World Meteorological Organization, Technical Note no. 143, WMO no. 415, 192 pp.

Svensson, C., Kundzewicz, Z. W., and Maurer, T. (2005). Trend detection in river flow series: 2. Floods and low-flow index series. Hydrological Sciences Journal, 50(5): 811-823.
}
