% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmr.R
\name{spear}
\alias{spear}
\title{Spearman's Rank Correlation Test}
\usage{
spear(x)
}
\arguments{
\item{x}{- Time series data vector}
}
\value{
Correlation coefficient - Spearman's Correlation coefficient value

Z-Tranformed Test Statistic value - Z-transform value to test significance \eqn{\rho(\sqrt{n-1})}
}
\description{
Spearman's Rank Correlation test by Lehmann (1975) and Sneyers (1990) is useful in detecting trends.
}
\details{
Spearman's Rank Correlation test by Lehmann (1975) and Sneyers (1990) is implemeted in this function.
}
\examples{
x<-c(Nile)
spear(x)

}
\references{
Lehmann, E. L. (1975). Nonparametrics: statistical methods based on ranks. Holden-Day, Inc., California, 457 pp.

Sneyers, R. (1990). On the statistical analysis of series of observations. World Meteorological Organization, Technical Note no. 143, WMO no. 415, 192 pp.
}
