% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkttest.R
\name{mkttest}
\alias{mkttest}
\title{Mann-Kendall Trend Test of Time series Data with out any modifications}
\usage{
mkttest(x)
}
\arguments{
\item{x}{- Time series data vector}
}
\value{
Z  - Mann- Kendall Z-statistic

sen's slope  - sen's slope

s  - Mann-Kendall 's'- statistic

Var(s) - Variance of 's'

P-value  - Mann-Kendall P-Value

Tau  - Mann-Kendall's Tau
}
\description{
Mann-Kendall trend test is a non-parametric trend test used to identify monotonic trends present in time series data.
}
\details{
Mann-Kendall trend test is a non-parametric trend tests which assuems no distribution of the data. Null hypothesis of the test is that there is no trend in the data and the alternate hypothesis is that the data represents a monotonic trend.
}
\examples{
x<-c(Nile)
mkttest(x)

}
\references{
Mann, H. B. (1945). Nonparametric Tests Against Trend. Econometrica, 13(3), 245–259. <doi:10.1017/CBO9781107415324.004>.

Kendall, M. (1975). Multivariate analysis. Charles Griffin. Londres. 0-85264-234-2.

sen, P. K. (1968). Estimates of the Regression Coefficient Based on Kendall’s Tau. Journal of the American statistical Association, 63(324), 1379. <doi:10.2307/2285891>.
}
