% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmr.R
\name{spear}
\alias{spear}
\title{Spearman's Rank Correlation Test}
\usage{
spear(x)
}
\arguments{
\item{x}{- Time series data vector}
}
\value{
Correlation coefficient - Spearman's Correlation coefficient value

Z-Tranformed Test Statistic value - Z-Transform value to test significance (Rho*(sqrt(n-1)))
}
\description{
Spearman's Rank Correlation test by Sneyers (1990) is an useful in detecting trends.
}
\examples{
x<-c(Nile)
spear(x)

}
\references{
Yue, S., & Wang, C. Y. (2004). The Mann-Kendall test modified by effective sample size to detect trend in serially correlated hydrological series. Water Resources Management, 18(3), 201–218. <doi:10.1023/B:WARM.0000043140.61082.60>.
}
